Renamed all delegate exceptions so that they are easier to identify.
